added the engine start/stop rendering. With this setup, your renderer will only proceed with the rendering process if the running flag is true. This allows the Engine to control the start and stop of the rendering.
However, note that simply not drawing when running is false won't stop the MTKView from attempting to draw frames. It simply means that when draw(in:) is called, the renderer will immediately return without doing any rendering work. If you want to stop the MTKView from trying to draw entirely, you would need to remove it from the view hierarchy or hide it.

Also, the running flag in Engine could be marked with @Published if you want changes to it to trigger UI updates in SwiftUI, though this doesn't seem necessary from what you've described.

Make sure your Renderer object stays in memory (e.g., it's retained by a @StateObject or similar in SwiftUI) and that you don't have multiple Engine objects unless that's intentional for your design.